Only look for shared creds if necessary
WHY:
Currently in `client.rb` there a number of places are checked for creds.
A local variable representing the shared creds file in
`~/.aws/credentials` is created before its known whether that option is
even necessary (i.e. before all the switching begins). This was causing
the client to error out on my machine because I use `ENV` variables and
didn't have a `[default]` profile in my shared creds file, which is what
this was trying to find (see #258).

Rather than create this variable when it's not even clear that we need
it, we can wait to use `Aws::SharedCredentials` when the control flow
reaches that point and more highly prioritized options have been ruled
out.

This also simplifies a number of specs in `client_spec.rb` since the
call to `SharedCredentials` doesn't need to be stubbed out every time.