Account for missing creds file or profile name
Aws::SharedCredentials raises an exception if you try to initialize
it with a profile name that doesn't exist. If you initialize with no
profile name, it will look for a [default] profile.

This change allows that step in the credentials check to avoid an error if, for example, the
user is intending to authenticate with an IAM instance profile and
has no ~/.aws/credentials or, in other cases, for whatever reason has
the credentials file but no [default] profile and hasn't supplied an
alternative.