add ability to disable AWS-managed key pairs
If :aws_ssh_key_id is nil, continue to perform the key auto-generation.

If :aws_ssh_key_id is explicitly set to "_disabled", do not generate a
key. This means that the kitchen configuration must be more explicit
about how connections will be authenticated. For example, when an
environment has disallowed AWS-managed keys, a key could be named in the
driver's `user_data` and the private key specified in the transport.

* moved the gnarly setup/expectations for #create_key to a separate
  describe block so that the expectations around key auto-generation
  or enable/disable are more clear

Co-authored-by: Robb Kidd <robb@thekidds.org>
Co-authored-by: Seth Thomas <sthomas@chef.io>
Signed-off-by: Robb Kidd <robb@thekidds.org>
Signed-off-by: Seth Thomas <sthomas@chef.io>