Use defaults when creating spot instances
This switches back to creating one-time requests and terminate on
interrupt, which is the default behavior. THat is needed so that
the block_duration_minutes feature will work with the API. The
non-default setting swere being used in an attempt to use the
valid_until setting, but we are already honoring the spot_wait
setting by using it in the retry.

Signed-off-by: Clinton Wolfe <clintoncwolfe@gmail.com>