feat: Tag security groups and key pairs
Kitchen-ec2 currently tags only some of the resources this creates. This
presented me with problems when I was trying to write a tag-restricted
IAM policy for an instance cleanup tool.

This patch causes kitchen-ec2 to tag any security group or ssh key pair
for which it owns the entire lifecycle.

Signed-off-by: Andrew Bobulsky <rulerof@gmail.com>